+++ pandas_datareader/google/quotes.py
@@ -1,0 +1,2 @@
+def get_quote_google(symbols):
+    raise NotImplementedError("Google Finance doesn't have this functionality")
